requirements.txt
absl-py==2.0.0
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.103.1
filelock==3.12.4
flatbuffers==23.5.26
fsspec==2023.9.1
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httptools==0.6.0
huggingface-hub==0.17.2
idna==3.4
jax==0.4.16
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
ml-dtypes==0.3.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.2
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.14.1
